♻️ 전체 코드수정본(통합)
