+++ iMacDonald/iMacDonald/SceneDelegate.swift
@@ -23,7 +23,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         // window 에게 루트 뷰 지정.
         // 이 메서드를 반드시 작성해줘야 윈도우가 활성화 됨.
-        window.rootViewController = TestViewController()
+        window.rootViewController = MainView()
         window.makeKeyAndVisible()
         self.window = window // 위의 var window 선언(넣어주겠다는 뜻)
     }
